requirements.txt
# local package
-e .

# external requirements
click
Sphinx
#coverage
#awscli
flake8
python-dotenv>=0.5.1
psycopg2-binary
SQLAlchemy
pandas
jupyterlab
ydata-profiling
ipywidgets
pyarrow
ratelimiter
nltk
svgling
langchain
langchain-google-genai
faiss-gpu

# backwards compatibility
pathlib2